@@ -10,7 +10,7 @@ description: A new Flutter project.
 version: 2.0.0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,7 +22,7 @@ dependencies:
   bloc_provider: ^1.0.0
   cupertino_icons: ^1.0.5
   cloud_firestore: ^3.1.9
-  firebase_admob: ^0.11.2
+  google_mobile_ads: ^1.0.0
   firebase_analytics: ^9.1.1
   firebase_core: ^1.12.0
   firebase_crashlytics: ^2.5.2
@@ -41,10 +41,20 @@ dependencies:
   rxdart: ^0.27.7
   shared_preferences: ^2.0.7
   url_launcher: ^6.0.10
+  collection: ^1.15.0-nullsafety.4
 
-#dependency_overrides:
+dependency_overrides:
   #intl: ^0.17.0
-
+  firebase_messaging_platform_interface: 3.1.6
+  firebase_storage_platform_interface: 4.0.14
+  firebase_crashlytics_platform_interface: 3.1.13
+  cloud_functions_platform_interface: 5.0.21
+  cloud_firestore_platform_interface: 5.4.13
+  firebase_auth_platform_interface: 6.1.11
+  firebase_database_platform_interface: 0.2.0+5
+  firebase_performance_platform_interface: 0.1.0+4
+  firebase_analytics_platform_interface: 3.0.5
+  firebase_remote_config_platform_interface: 1.0.5
 dev_dependencies:
   flutter_test:
     sdk: flutter
