@@ -1,19 +1,19 @@
 import 'package:bloc_provider/bloc_provider.dart';
+import 'package:collection/collection.dart' show IterableNullableExtension;
 import 'package:disclosure_app_fl/bloc/bloc.dart';
 import 'package:disclosure_app_fl/bloc/company_disclosure_bloc2.dart';
 import 'package:disclosure_app_fl/models/company-settlement.dart';
 import 'package:disclosure_app_fl/models/company.dart';
 import 'package:disclosure_app_fl/models/edinet.dart';
-import 'package:disclosure_app_fl/utils/admob.dart';
 import 'package:disclosure_app_fl/widgets/disclosure_list_item.dart';
 import 'package:disclosure_app_fl/widgets/edinet_streaming.dart';
 import 'package:disclosure_app_fl/widgets/no_disclosures.dart';
-import 'package:firebase_admob/firebase_admob.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 class DisclosureCompanyScreen extends StatefulWidget {
-  final Company company;
+  final Company? company;
 
   DisclosureCompanyScreen({this.company});
 
@@ -23,29 +23,29 @@ class DisclosureCompanyScreen extends StatefulWidget {
 }
 
 class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
-  final Company company;
-  BannerAd banner;
-  int tabLength;
-  CompanyDisclosureBloc2 bloc2;
+  final Company? company;
+  // BannerAd banner;
+  int? tabLength;
+  CompanyDisclosureBloc2? bloc2;
 
   @override
   initState() {
     super.initState();
-    banner = showBanner("ca-app-pub-5131663294295156/4027309882");
+    // banner = showBanner("ca-app-pub-5131663294295156/4027309882");
     final appBloc = BlocProvider.of<AppBloc>(context);
-    if (company.edinetCode != '') {
+    if (company!.edinetCode != '') {
       tabLength = 2;
     } else {
       tabLength = 1;
     }
     this.bloc2 = CompanyDisclosureBloc2(
-      this.company,
+      this.company!,
       companies: appBloc.companyMap$,
       user$: appBloc.user$,
     );
-    this.bloc2.disclosure.reload.add(null);
-    if (company.edinetCode != '') {
-      this.bloc2.edinet.reload.add(null);
+    this.bloc2!.disclosure!.reload.add(null);
+    if (company!.edinetCode != '') {
+      this.bloc2!.edinet!.reload.add(null);
     }
   }
 
@@ -56,15 +56,15 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
       child: (snapshot.length == 0
           ? const NoDisclosures()
           : CustomScrollView(slivers: <Widget>[
-              StreamBuilder<CompanySettlement>(
-                stream: bloc2.companySettlement$,
+              StreamBuilder<CompanySettlement?>(
+                stream: bloc2!.companySettlement$,
                 builder: (context, snapshot) {
                   return SliverList(
                     delegate: SliverChildListDelegate(
                       (!snapshot.hasData ||
                               snapshot.data == null ||
-                              snapshot.data.schedule == null ||
-                              snapshot.data.schedule
+                              snapshot.data!.schedule == null ||
+                              snapshot.data!.schedule!
                                   .add(Duration(days: 1))
                                   .isBefore(DateTime.now()))
                           ? []
@@ -78,7 +78,7 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
                                         color:
                                             Theme.of(context).backgroundColor,
                                       ),
-                                      title: Text(snapshot.data.toMessage()),
+                                      title: Text(snapshot.data!.toMessage()),
                                     )
                                   ],
                                 ),
@@ -92,10 +92,10 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
               SliverList(
                 delegate: SliverChildBuilderDelegate(
                   (context, index) => DisclosureListItem(
-                        item: snapshot[index],
-                        showDate: true,
-                        key: Key(snapshot[index]['document']),
-                      ),
+                    item: snapshot[index],
+                    showDate: true,
+                    key: Key(snapshot[index]['document']),
+                  ),
                   childCount: snapshot.length,
                 ),
               ),
@@ -104,29 +104,30 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
                   padding: const EdgeInsets.all(8.0),
                   child: RaisedButton(
                     child: StreamBuilder<bool>(
-                        stream: bloc2.disclosure.isLoading$,
+                        stream: bloc2!.disclosure!.isLoading$,
                         builder: (context, snapshot) {
                           return snapshot.data == false
                               ? Text('更に読み込む')
                               : CircularProgressIndicator();
                         }),
                     onPressed: () {
-                      this.bloc2.disclosure?.loadNext?.add(snapshot.last);
+                      this.bloc2!.disclosure?.loadNext?.add(snapshot.last);
                     },
                   ),
                 ),
               )
             ])),
       onRefresh: () {
-        bloc2.disclosure?.reload?.add(this.company.code);
-        return bloc2.disclosure?.isLoading$?.where((e) => e)?.first;
+        bloc2!.disclosure?.reload?.add(this.company!.code);
+        return bloc2!.disclosure?.isLoading$?.where((e) => e).first
+            as Future<void>;
       },
     );
   }
 
   Widget _buildBody(BuildContext context) {
-    return StreamBuilder<List<DocumentSnapshot>>(
-      stream: bloc2.disclosure.data$,
+    return StreamBuilder<List<DocumentSnapshot>?>(
+      stream: bloc2!.disclosure!.data$,
       builder: (context, snapshot) {
         if (!snapshot.hasData || snapshot.data == null) {
           return Column(
@@ -135,7 +136,7 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
             ],
           );
         }
-        return _buildList(context, snapshot.data);
+        return _buildList(context, snapshot.data!);
       },
     );
   }
@@ -146,21 +147,21 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
     final mediaQuery = MediaQuery.of(context);
 
     return DefaultTabController(
-      length: this.tabLength,
+      length: this.tabLength!,
       child: Scaffold(
         appBar: AppBar(
-          title: Text("${this.company.name} (${this.company.code})"),
+          title: Text("${this.company!.name} (${this.company!.code})"),
           actions: <Widget>[
             StreamBuilder<List<Company>>(
               stream: appBloc.favoritesWithName$,
               builder: (context, snapshot) {
                 final isFavorite = snapshot.hasData &&
-                    snapshot.data.any((fav) => fav.code == this.company.code);
+                    snapshot.data!.any((fav) => fav.code == this.company!.code);
                 return IconButton(
                   icon: Icon(isFavorite ? Icons.star : Icons.star_border),
                   tooltip: 'お気に入り',
                   onPressed: () {
-                    appBloc.switchFavorite.add(this.company.code);
+                    appBloc.switchFavorite.add(this.company!.code);
                     Scaffold.of(context).showSnackBar(SnackBar(
                       content:
                           Text(isFavorite ? 'お気に入りを解除しました' : 'お気に入りに追加しました'),
@@ -194,29 +195,29 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
             Tab(
               text: 'TDNET',
             ),
-            if (this.company.edinetCode != '')
+            if (this.company!.edinetCode != '')
               Tab(
                 text: 'EDINET',
               ),
           ]),
         ),
         body: TabBarView(
-            children: <Widget>[
+            children: <Widget?>[
           _buildBody(context),
           this.tabLength == 2 ? _buildEdinetList(context) : null,
-        ].where((e) => e != null).toList()),
-        persistentFooterButtons: <Widget>[
-          SizedBox(
-            height: getSmartBannerHeight(mediaQuery) - 16.0,
-          )
-        ],
+        ].whereNotNull().toList()),
+        // persistentFooterButtons: <Widget>[
+        //   SizedBox(
+        //     height: getSmartBannerHeight(mediaQuery) - 16.0,
+        //   )
+        // ],
       ),
     );
   }
 
   Widget _buildEdinetList(BuildContext context) {
-    return StreamBuilder<List<Edinet>>(
-        stream: bloc2.edinet.data$,
+    return StreamBuilder<List<Edinet>?>(
+        stream: bloc2!.edinet!.data$,
         builder: (context, snapshot) {
           return RefreshIndicator(
             child: Builder(builder: (context) {
@@ -227,7 +228,7 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
                   ],
                 );
               }
-              if (snapshot.data.length == 0) {
+              if (snapshot.data!.length == 0) {
                 return Container(
                   alignment: Alignment.center,
                   child: Column(
@@ -241,12 +242,12 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
               }
               return ListView.builder(
                 itemBuilder: (context, idx) {
-                  return idx == snapshot.data.length
+                  return idx == snapshot.data!.length
                       ? Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: RaisedButton(
                             child: StreamBuilder<bool>(
-                              stream: bloc2.edinet.isLoading$,
+                              stream: bloc2!.edinet!.isLoading$,
                               builder: (context, snapshot) {
                                 return snapshot.data == false
                                     ? Text('更に読み込む')
@@ -255,24 +256,24 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
                             ),
                             onPressed: () {
                               this
-                                  .bloc2
+                                  .bloc2!
                                   .edinet
                                   ?.loadNext
-                                  ?.add(snapshot.data.last);
+                                  ?.add(snapshot.data!.last);
                             },
                           ),
                         )
                       : EdinetListItem(
-                          edinet: snapshot.data[idx],
+                          edinet: snapshot.data![idx],
                           showDate: true,
                         );
                 },
-                itemCount: snapshot.data.length + 1,
+                itemCount: snapshot.data!.length + 1,
               );
             }),
             onRefresh: () {
-              bloc2.edinet.reload.add(null);
-              return bloc2.edinet.data$.first;
+              bloc2!.edinet!.reload.add(null);
+              return bloc2!.edinet!.data$.first;
             },
           );
         });
@@ -280,7 +281,7 @@ class _DisclosureCompanyScreenState extends State<DisclosureCompanyScreen> {
 
   @override
   void dispose() {
-    banner?.dispose();
+    // banner?.dispose();
     bloc2?.dispose();
     super.dispose();
   }
