@@ -14,28 +14,34 @@ final dateFormatter = DateFormat("yyyy-MM-dd");
 class EdinetBloc extends Bloc {
   final path = 'edinets';
 
-  final _edinet$ = BehaviorSubject<List<Edinet>>();
+  final _edinet$ = BehaviorSubject<List<Edinet>?>();
   // final _filter$ = BehaviorSubject<String>(seedValue: '');
 
-  ValueStream<List<Edinet>> get edinet$ => _edinet$.stream;
+  ValueStream<List<Edinet>?> get edinet$ => _edinet$.stream;
   // ValueObservable<String> get filter$ => _filter$;
   // Sink<String> get filterController => _filter$.sink;
 
-  Stream<List<Edinet>> edinetStream(AppBloc bloc) {
+  Stream<List<Edinet>?> edinetStream(AppBloc bloc) {
     final _edinets =
         bloc.user$.switchMap((_) => bloc.edinetDate$).switchMap((date) {
       print("***date*** $date");
       final start = dateFormatter.format(date);
       final end = dateFormatter.format(date.add(Duration(days: 1)));
-      return FirebaseFirestore.instance.collection(path).where('seq', isGreaterThanOrEqualTo: start).where('seq', isLessThan: end).orderBy('seq', descending: true).snapshots().startWith(null);
+      return (FirebaseFirestore.instance
+              .collection(path)
+              .where('seq', isGreaterThanOrEqualTo: start)
+              .where('seq', isLessThan: end)
+              .orderBy('seq', descending: true)
+              .snapshots() as Stream<QuerySnapshot<Map<String, dynamic>>?>)
+          .startWith(null);
     }).map((doc) {
       return doc?.docs;
     });
 
     final _mappedEdinets = Rx.combineLatest2<
-        List<DocumentSnapshot>,
+        List<DocumentSnapshot>?,
         Map<String, Company>,
-        Iterable<Edinet>>(_edinets, bloc.companyMap$, (edinets, companies) {
+        Iterable<Edinet>?>(_edinets, bloc.companyMap$, (edinets, companies) {
       if (edinets == null) {
         return null;
       }
@@ -46,18 +52,17 @@ class EdinetBloc extends Bloc {
       });
     });
 
-    final _favorite$ =
-        Rx.combineLatest2<List<Company>, bool, List<Company>>(
-            bloc.favoritesWithName$, bloc.edinetShowOnlyFavorite$,
-            (_favorites, _favoriteOnly) {
+    final _favorite$ = Rx.combineLatest2<List<Company>, bool, List<Company>>(
+        bloc.favoritesWithName$, bloc.edinetShowOnlyFavorite$,
+        (_favorites, _favoriteOnly) {
       if (_favoriteOnly == false) {
         return [];
       }
       return _favorites;
     });
 
-    return Rx.combineLatest3<Iterable<Edinet>, String, List<Company>,
-            List<Edinet>>(_mappedEdinets, bloc.edinetFilter$, _favorite$,
+    return Rx.combineLatest3<Iterable<Edinet>?, String, List<Company>,
+            List<Edinet>?>(_mappedEdinets, bloc.edinetFilter$, _favorite$,
         (edinets, filter, favorites) {
       if (edinets == null) {
         return null;
